requirements.txt
asyncio==3.4.3
discord.py==1.5.1